Cline/MCP/agentics-mcp/requirements.txt
fastmcp==2.11.2
fastapi>=0.116.1
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pyyaml>=6.0
